+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Middleware for parsing AuthCode and state params
+
 ## [0.1.0] - 2022-01-04
 
 ### Added